Read-in and merge BEDPE Interactions by Genomic distance with DBSCAN

Given a list of BEDPE files, a resolution, and a selector column, this function will use dbscan to merge loops by genomic distance.

mergeBedpe(
  bedpeFiles,
  res = 5000,
  selectCol = 12,
  dist_method = "manhattan",
  minPts = 2
)

Arguments

bedpeFiles
Character vector of BEDPE file paths to be merged.

res
Integer - resolution in base-pairs of all BEDPE interactions. Used to determine the epsilon value for dbscan() (i.e. eps = res*2).

selectCol
Integer or character string denoting the column to be used to select clustered interactions.

dist_method
Character string - distance measure used in the dist function. For available methods see ?dist().

minPts
Integer - minimum number of interactions to form a DBSCAN cluster.

Value

Returns a data table of selected BEDPE interactions